Cleanup UI
Uniformize textbox and button sizes.
Realign components to provide a cleaner feel.
It helped gain a few pixels here there.
Correct a few wordings so they are consistant between the main window and the test window.
Removed the option expander from the test windows since it is on eline high and it contains one line of options.